requirements.txt
Flask==1.1.1
Flask-JWT-Extended==3.22.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
passlib==1.7.1
jsonpatch==1.24
Pillow==6.1.0
pytest==5.1.2
